refactor: migrate to many-to-many note-tag relation and rename entities
- Rename Note/Tag to NoteEntity/TagEntity to clarify database layer
- Relocate NoteUi/TagUi to domain package for better architecture separation
- Implement many-to-many relationship using NoteTagLink associative table
- Add Room Migration (v2 to v3) to handle table recreation and data transfer
- Update DAOs and Repositories to support NoteWithTags and TagWithNotes relations aiming to build, but not work properly
- Update UI components to reflect model and relationship changes aiming to build, but not work properly
- Update color schemes and typography

BREAKING CHANGE: Database schema updated from v2 to v3. Existing notes_table and tags_table are replaced by notes and tags.

+++ app/src/main/java/br/com/arml/insights/ui/screen/note/NoteEvent.kt
@@ -1,6 +1,6 @@
 package br.com.arml.insights.ui.screen.note
 
-import br.com.arml.insights.model.entity.NoteUi
+import br.com.arml.insights.model.domain.NoteUi
 import br.com.arml.insights.ui.screen.common.Reducer
 import br.com.arml.insights.utils.data.SearchNoteCategory
 
@@ -16,7 +16,7 @@ sealed class NoteEvent: Reducer.ViewEvent{
         val searchNoteCategory: SearchNoteCategory
     ): NoteEvent()
     data object OnFetchTags: NoteEvent()
-    data class OnSelectNewTag(val tagId: Int): NoteEvent()
+    data class OnSelectNewTag(val tagId: Long): NoteEvent()
 
     /* Actions to handle frontend events */
     data class OnClickToOpenSheet(
